Move sam2 notebooks to separate places
