+++ notebooks/README.md
@@ -37,8 +37,8 @@
 - [Sound Generation with AudioLDM2 and OpenVINO™](./sound-generation-audioldm2/sound-generation-audioldm2.ipynb)
 - [SoftVC VITS Singing Voice Conversion and OpenVINO™](./softvc-voice-conversion/softvc-voice-conversion.ipynb)
 - [Object masks from prompts with SAM and OpenVINO](./segment-anything/segment-anything.ipynb)
-- [Object masks from prompts with SAM2 and OpenVINO](./segment-anything/segment-anything-2-video.ipynb)
-- [Object masks from prompts with SAM2 and OpenVINO for Images](./segment-anything/segment-anything-2-image.ipynb)
+- [Video segmentation with SAM2 and OpenVINO™](./sam2-video-segmentation/segment-anything-2-video.ipynb)
+- [Image segmentation with SAM2 and OpenVINO™](./sam2-image-segmentation/segment-anything-2-image.ipynb)
 - [Single step image generation using SDXL-turbo and OpenVINO](./sdxl-turbo/sdxl-turbo.ipynb)
 - [Visual-language assistant with Qwen2VL and OpenVINO](./qwen2-vl/qwen2-vl.ipynb)
 - [Audio-language assistant with Qwen2Audio and OpenVINO](./qwen2-audio/qwen2-audio.ipynb)
@@ -204,8 +204,8 @@
 - [One Step Sketch to Image translation with pix2pix-turbo and OpenVINO](./sketch-to-image-pix2pix-turbo/sketch-to-image-pix2pix-turbo.ipynb)
 - [Zero-shot Image Classification with SigLIP](./siglip-zero-shot-image-classification/siglip-zero-shot-image-classification.ipynb)
 - [Object masks from prompts with SAM and OpenVINO](./segment-anything/segment-anything.ipynb)
-- [Object masks from prompts with SAM2 and OpenVINO](./segment-anything/segment-anything-2-video.ipynb)
-- [Object masks from prompts with SAM2 and OpenVINO for Images](./segment-anything/segment-anything-2-image.ipynb)
+- [Image segmentation with SAM2 and OpenVINO™](./sam2-image-segmentation/segment-anything-2-image.ipynb)
+- [Video segmentation with SAM2 and OpenVINO™](./sam2-video-segmentation/segment-anything-2-video.ipynb)
 - [Single step image generation using SDXL-turbo and OpenVINO](./sdxl-turbo/sdxl-turbo.ipynb)
 - [Text-to-Video retrieval with S3D MIL-NCE and OpenVINO](./s3d-mil-nce-text-to-video-retrieval/s3d-mil-nce-text-to-video-retrieval.ipynb)
 - [Background removal with RMBG v1.4 and OpenVINO](./rmbg-background-removal/rmbg-background-removal.ipynb)
