+++ notebooks/segment-anything/README.md
@@ -3,20 +3,16 @@
 
 Segmentation - identifying which image pixels belong to an object - is a core task in computer vision and is used in a broad array of applications, from analyzing scientific imagery to editing photos. But creating an accurate segmentation model for specific tasks typically requires highly specialized work by technical experts with access to AI training infrastructure and large volumes of carefully annotated in-domain data. Reducing the need for task-specific modeling expertise, training compute, and custom data annotation for image segmentation is the main goal of the [Segment Anything](https://arxiv.org/abs/2304.02643) project.
 
-The Segment Anything Model (SAM) predicts object masks given prompts that indicate the desired object. SAM has learned a general notion of what objects are, and it can generate masks for any object in any image or any video, even including objects and image types that it had not encountered during training. SAM is general enough to cover a broad set of use cases and can be used out of the box on new image “domains” (e.g. underwater photos, MRI or cell microscopy) without requiring additional training (a capability often referred to as zero-shot transfer). 
+The [Segment Anything Model (SAM)](https://github.com/facebookresearch/segment-anything) predicts object masks given prompts that indicate the desired object. SAM has learned a general notion of what objects are, and it can generate masks for any object in any image or any video, even including objects and image types that it had not encountered during training. SAM is general enough to cover a broad set of use cases and can be used out of the box on new image “domains” (e.g. underwater photos, MRI or cell microscopy) without requiring additional training (a capability often referred to as zero-shot transfer). 
 
 Previously, to solve any kind of segmentation problem, there were two classes of approaches. The first, interactive segmentation, allowed for segmenting any class of object but required a person to guide the method by iterative refining a mask. The second, automatic segmentation, allowed for segmentation of specific object categories defined ahead of time (e.g., cats or chairs) but required substantial amounts of manually annotated objects to train (e.g., thousands or even tens of thousands of examples of segmented cats), along with the compute resources and technical expertise to train the segmentation model. Neither approach provided a general, fully automatic approach to segmentation.
 
 Segment Anything Model is a generalization of these two classes of approaches. It is a single model that can easily perform both interactive segmentation and automatic segmentation.
+This notebook shows an example of how to convert and use Segment Anything Model in OpenVINO format, allowing it to run on a variety of platforms that support an OpenVINO.
 
-There two version of models [SAM](https://github.com/facebookresearch/segment-anything) and [SAM2](https://github.com/facebookresearch/segment-anything-2). SAM2 is expand SAM to video by considering images as a video with a single frame. Examples of how to convert and use these models in OpenVINO format, allowing it to run on a variety of platforms that support an OpenVINO are presented in next notebooks:
-- [`segment-anything.ipynb`](./segment-anything.ipynb)
-- [`segment-anything-2-image.ipynb`](./segment-anything-2-image.ipynb)
-- [`segment-anything-2-video.ipynb`](./segment-anything-2-video.ipynb)
+The notebook demonstrates how to work with model in 2 modes:
 
-The notebooks demonstrates how to work with model in 2 modes:
-
-* Interactive segmentation mode: in this demonstration you can upload image/video and specify point related to object using [Gradio](https://gradio.app/) interface and as the result you get segmentation mask for specified point.
+* Interactive segmentation mode: in this demonstration you can upload image and specify point related to object using [Gradio](https://gradio.app/) interface and as the result you get segmentation mask for specified point.
 The following image shows an example of the input text and the corresponding predicted image.
 ![demo](https://user-images.githubusercontent.com/29454499/231464914-bd2a683c-28b2-44d4-960e-dce3e3ddebc3.png)
 
@@ -24,17 +20,13 @@ The following image shows an example of the input text and the corresponding pre
 
 ![demo2](https://user-images.githubusercontent.com/29454499/231468849-1cd11e68-21e2-44ed-8088-b792ef50c32d.png)
 
-
 ## Notebook Contents
-
 This notebook shows an example of how to convert and use Segment Anything Model using OpenVINO
-
 Notebook contains the following steps:
 1. Convert PyTorch models to OpenVINO format.
 2. Run OpenVINO model in interactive segmentation mode.
 3. Run OpenVINO model in automatic mask generation mode.
 4. Run NNCF post-training optimization pipeline to compress the encoder of SAM
-5. For SAM2 video: Convert PyTorch models to OpenVINO format and run model in interactive segmentation mode with video.
 
 
 ## Installation Instructions
